+++ lib/puppet_x/cf_db.rb
@@ -1,4 +1,12 @@
 
+# Done this way due to some weird behavior in tests also ignoring $LOAD_PATH
+begin
+    require File.expand_path( '../cf_system', __FILE__ )
+rescue LoadError
+    require File.expand_path( '../../../../cfsystem/lib/puppet_x/cf_system', __FILE__ )
+end
+
+
 module PuppetX::CfDb
     CFDB_TYPES = [
         'MySQL',
@@ -7,6 +15,7 @@ module PuppetX::CfDb
     ROOT_PASS_LEN = 24
     BASE_DIR = File.expand_path('../', __FILE__)
     SECURE_PORT_OFFSET = 50
+    ACCESS_CHECK_TOOL = "#{PuppetX::CfSystem::CUSTOM_BIN_DIR}/cfdb_access_checker"
 
     #---
     require "#{BASE_DIR}/cf_db/provider_base"
@@ -14,7 +23,6 @@ module PuppetX::CfDb
     CFDB_TYPES.each do |t|
         t = t.downcase()
         require "#{BASE_DIR}/cf_db/#{t}"
-        require "#{BASE_DIR}/cf_db/#{t}/access"
         require "#{BASE_DIR}/cf_db/#{t}/database"
         require "#{BASE_DIR}/cf_db/#{t}/instance"
         require "#{BASE_DIR}/cf_db/#{t}/role"